폭발은 예술! 로켓런쳐!
no michael, you don't understand. the US Defense Budget in 2011 was
664.84 billion dollars. That's enough (counting saving money from a
dozen donuts only being 7 dollars) to buy approximately
1,140,000,000,000 donuts. That's One Trillion donuts. According to my
research, a donut is the size of a hockey puck. A hockey puck is one
inch thick and 3 inches in diameter. That's around 2,555 square miles
(while that doesn't sound big, it's actually around 124,000 football
fields, coated in a solid layer of donuts). This is around the size of
Delaware, or New York City 5 times over. If you stacked them on top of
each other, you could make a spire of donuts that stretches 17,992,424
miles into the sky. The moon is 233,014 miles from Earth. That could
almost reach Venus, when it's at its closest ever to Earth.
But I'm not done. The volume of a donut is around 7 cubic inches (note
I'm assuming these are boston creme donuts, the best kind, so that
there's no hole to calculate around). The total volume of all these
donuts is 7,980,000,000,000 ci, or a solid wall (3 feet thick, 16 feet
tall) 1,154,514,000 inches long, or 18,221 miles. That's almost enough
to wrap around the Earth.
All of these donuts weighs 6,270,000,000,000 ounces, or 48,984,375,000
gallons. That is half of all the oil still remaining in the gulf after
the deepwater spill. That's enough to fill around a million swimming
pools. You could have a cube 150 ft long on each side, and with donuts
still left over.
Each donut has 240 Calories. That's enough to feed 1,666,000,000 people
for three months solid, or more than the entire population of africa.
Those calories can be converted into 4,121,236,800,000 kilowatt hours.
This is enough to power 100,000,000 lightbulbs running for 10 hours, or
keep 100 houses running for 2,000 years. It would produce the power
equal to 273,610 tons of TNT, or 200 Hiroshimas. This would be an
explosion that would completely wipe out 100 square miles, with a total
explosion range of 400 miles. This is enough to turn NYC into dust.